refactor of controller classes
