+++ src/main/java/com/diplom/sptor/web/RepairController.java
@@ -22,52 +22,36 @@ import java.util.*;
  * Created by user on 02.02.2016.
  */
 @Controller
-@RequestMapping(value = "/rep")
+@RequestMapping(value = "/repair")
 public class RepairController {
-   // private EquipmentService EquipmentService;
-   // private SubdivisionService SubdivisionService;
-   // private RepairSheetService RepairSheetService;
-   // private TypeOfMaintenanceService TypeOfMaintenanceService;
-   // private StatusService StatusService;
-   // private StatusOfEqService StatusOfEqService;
-
-    @Autowired
-    UserService userService;
 
     @Autowired
     SpareService spareService;
 
     @Autowired
-    ComponentManager componentManager;
-  //@Autowired(required=true)
-  //public void setEquipmentService(EquipmentService EquipmentService){
-  //    this.EquipmentService = EquipmentService;
-  //}
+    ComponentService componentService;
 
-  //@Autowired(required=true)
-  //public void setSubdivisionService(SubdivisionService SubdivisionService){
-  //    this.SubdivisionService = SubdivisionService;
-  //}
+    @Autowired
+    EquipmentService equipmentService;
 
-  //@Autowired(required=true)
-  //public void setRepairSheetService(RepairSheetService RepairSheetService){
-  //    this.RepairSheetService = RepairSheetService;
-  //}
+    @Autowired
+    UserService userService;
 
-   // @Autowired(required=true)
-   // public void setTypeOfMaintenanceService(TypeOfMaintenanceService TypeOfMaintenanceService) {
-   //     this.TypeOfMaintenanceService = TypeOfMaintenanceService;
-   // }
-//
-   // @Autowired(required=true)
-   // public void setStatusService(StatusService StatusService) {
-   //     this.StatusService = StatusService;
-   // }
-//
-   // @Autowired(required=true)
-   // public void setStatusOfEqService(StatusOfEqService StatusOfEqService) {
-   //     this.StatusOfEqService = StatusOfEqService;
-   // }
+    @Autowired
+    SubdivisionService subdivisionService;
+
+    @Autowired
+    TypeOfMaintenanceService typeOfMaintenanceService;
+
+    @Autowired
+    StatusService statusService;
+
+    @Autowired
+    RepairSheetService repairSheetService;
+
+    @Autowired
+    StatusOfEqService statusOfEqService;
+
 
     private String getPrincipal(){
         String userName = null;
@@ -82,29 +66,31 @@ public class RepairController {
     }
 
     @ApiOperation(value = "getEquipments", notes = "Get all equipments")
-    @RequestMapping(value = "/", method = RequestMethod.GET,
+    @RequestMapping(method = RequestMethod.GET,
             produces = "application/json")
     public String RepairStart(Model model) {
-       // model.addAttribute("repairSheet", new RepairSheet());
-       // model.addAttribute("listRepair", this.RepairSheetService.listRepairSheets());
-       // model.addAttribute("status_one", this.RepairSheetService.getRepairSheetByStatus(this.StatusService.getStatusById(1)).size());
-       // model.addAttribute("status_two", this.RepairSheetService.getRepairSheetByStatus(this.StatusService.getStatusById(2)).size() + this.RepairSheetService.getRepairSheetByStatus(this.StatusService.getStatusById(3)).size());
-       // model.addAttribute("status_three", this.RepairSheetService.getRepairSheetByStatus(this.StatusService.getStatusById(4)).size());
-       // model.addAttribute("status_four", this.RepairSheetService.getRepairSheetByStatus(this.StatusService.getStatusById(5)).size());
-       // model.addAttribute("status_all", this.RepairSheetService.listRepairSheets().size());
-       // model.addAttribute("subdivisions", this.SubdivisionService.listSubdivisions());
-        model.addAttribute("components", componentManager.getAllComponents());
-        model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " VAAN " +
+        model.addAttribute("repairSheet", new RepairSheet());
+        model.addAttribute("listRepair", repairSheetService.getAllRepairSheets());
+        model.addAttribute("status_one", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(1)).size());
+        model.addAttribute("status_two", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(2)).size() + repairSheetService.getRepairSheetByStatus(statusService.getStatusById(3)).size());
+        model.addAttribute("status_three", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(4)).size());
+        model.addAttribute("status_four", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(5)).size());
+        model.addAttribute("status_all", repairSheetService.getAllRepairSheets().size());
+        model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
+        model.addAttribute("components", componentService.getAllComponents());
+        model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
                 userService.getUserBySso(getPrincipal()).getFirst_name());
-      //  model.addAttribute("listEquipments", this.EquipmentService.listEquipments());
-      //  model.addAttribute("listTypeOfMaintenance", this.TypeOfMaintenanceService.listType());
-      // Status status1 = StatusService.getStatusById(1);
-      // Status status2 = StatusService.getStatusById(2);
-      // model.addAttribute("active_req", this.RepairSheetService.getRepairSheetByStatus(status1).size());
-      // model.addAttribute("confirm_req", this.RepairSheetService.getRepairSheetByStatus(status2).size());
-      // model.addAttribute("listStatus", this.StatusService.listStatus());
+        model.addAttribute("listEquipments", equipmentService.getAllEquipment());
+        model.addAttribute("listTypeOfMaintenance", typeOfMaintenanceService.getAllTypes());
+        Status status1 = statusService.getStatusById(1);
+        Status status2 = statusService.getStatusById(2);
+        model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
+        model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
+        model.addAttribute("listStatus", repairSheetService.getAllRepairSheets());
         return "repair";
     }
+
+
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
     public String addRepairSheet(
@@ -120,22 +106,22 @@ public class RepairController {
         Date end = formatter.parse(start_date);
         DateTime dateTime = new DateTime(end);
         DateTimeFormatter dtf = DateTimeFormat.forPattern("mm-dd-yyyy");
-     //   DateTime end_date = dateTime.plusDays(this.TypeOfMaintenanceService.getTypeById(type_of_maintenance_id).getDuration());
-     //   Status status = this.StatusService.getStatusById(1);
+        DateTime end_date = dateTime.plusDays(typeOfMaintenanceService.getTypeById(type_of_maintenance_id).getDuration());
+        Status status = statusService.getStatusById(1);
         User resp_for_delivery = userService.getUserById(1);
         User resp_for_reception = userService.getUserById(1);
-     //   Equipment equipment = this.EquipmentService.getEquipmentById(equipment_id);
-        Components components1 = componentManager.getComponentById(components);
-      //  Subdivisions subdivisions = this.SubdivisionService.getSubdivisionById(subdivision_id);
-     //   TypeOfMaintenance typeOfMaintenance = this.TypeOfMaintenanceService.getTypeById(type_of_maintenance_id);
+        Equipment equipment = equipmentService.getEquipmentById(equipment_id);
+        Components components1 = componentService.getComponentById(components);
+        Subdivisions subdivisions = subdivisionService.getSubdivisionById(subdivision_id);
+        TypeOfMaintenance typeOfMaintenance = typeOfMaintenanceService.getTypeById(type_of_maintenance_id);
         int sheet_number = equipment_id + 33000;
         int warranty_period = type_of_maintenance_id;
-      // this.RepairSheetService.addRepairSheet(new RepairSheet(typeOfMaintenance, equipment,
-      //         components1, subdivisions, start, end_date.toDate(), sheet_number, warranty_period,
-      //         resp_for_delivery, resp_for_reception, description, status));
+        repairSheetService.addRepairSheet(new RepairSheet(typeOfMaintenance, equipment,
+                components1, subdivisions, start, end_date.toDate(), sheet_number, warranty_period,
+                resp_for_delivery, resp_for_reception, description, status));
 
         deleteComponentInStock(components1);
-      //  changeStatusOfEquipment(equipment, status, typeOfMaintenance);
+        changeStatusOfEquipment(equipment, status, typeOfMaintenance);
 
         return "redirect:/repair";
 
@@ -147,103 +133,102 @@ public class RepairController {
         //this.SpareService.updateSpare(spares);
     }
 
-  // public void changeStatusOfEquipment(Equipment equipment, Status status, TypeOfMaintenance typeOfMaintenance){
-  //     int type = typeOfMaintenance.getType_of_maintenance_id();
-  //     switch (status.getStatus_id()){
-  //         case 1:
-  //             if(type==1||type==2||type==6||type==7||type==8){
-  //                 equipment.setStatus(this.StatusOfEqService.getStatusById(2));
-  //                 this.EquipmentService.updateEquipment(equipment);
-  //             }
-  //             else if(type==3||type==4||type==5){
-  //                 equipment.setStatus(this.StatusOfEqService.getStatusById(3));
-  //                 this.EquipmentService.updateEquipment(equipment);
-  //             }
-  //             else{
-  //                 equipment.setStatus(this.StatusOfEqService.getStatusById(4));
-  //                 this.EquipmentService.updateEquipment(equipment);
-  //             }
-  //             break;
-  //         case 4:
-  //             equipment.setStatus(this.StatusOfEqService.getStatusById(1));
-  //             this.EquipmentService.updateEquipment(equipment);
-  //             break;
-  //         case 5:
-  //             equipment.setStatus(this.StatusOfEqService.getStatusById(1));
-  //             this.EquipmentService.updateEquipment(equipment);
-  //             break;
-  //     }
-  // }
+    public void changeStatusOfEquipment(Equipment equipment, Status status, TypeOfMaintenance typeOfMaintenance){
+        int type = typeOfMaintenance.getType_of_maintenance_id();
+        switch (status.getStatus_id()){
+            case 1:
+                if(type==1||type==2||type==6||type==7||type==8){
+                    equipment.setStatus(statusOfEqService.getStatusById(2));
+                    equipmentService.updateEquipment(equipment);
+                }
+                else if(type==3||type==4||type==5){
+                    equipment.setStatus(statusOfEqService.getStatusById(3));
+                    equipmentService.updateEquipment(equipment);
+                }
+                else{
+                    equipment.setStatus(statusOfEqService.getStatusById(4));
+                    equipmentService.updateEquipment(equipment);
+                }
+                break;
+            case 4:
+                equipment.setStatus(statusOfEqService.getStatusById(1));
+                equipmentService.updateEquipment(equipment);
+                break;
+            case 5:
+                equipment.setStatus(statusOfEqService.getStatusById(1));
+                equipmentService.updateEquipment(equipment);
+                break;
+        }
+    }
 
-  //  @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-  //  @RequestMapping(value = "/{repairId}", method = RequestMethod.GET,
-  //          produces = "application/json")
-  //  public @ResponseBody RepairSheet getRepairById(@PathVariable("repairId") int repairId, Model model) {
-  //      return this.RepairSheetService.getRepairSheetById(repairId);
-  //  }
+    @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
+    @RequestMapping(value = "/{repairId}", method = RequestMethod.GET,
+            produces = "application/json")
+    public @ResponseBody RepairSheet getRepairById(@PathVariable("repairId") int repairId, Model model) {
+        return repairSheetService.getRepairSheetById(repairId);
+    }
 
-  // @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-  // @RequestMapping(value = "/equipment/{equipmentId}", method = RequestMethod.GET,
-  //         produces = "application/json")
-  // public @ResponseBody List<RepairSheet> getRepairByEquipmentId(@PathVariable("equipmentId") int equipmentId, Model model) {
-  //     List<RepairSheet> repairSheets = this.RepairSheetService.getRepairSheetByEquipment(this.EquipmentService.getEquipmentById(equipmentId));
-  //     Iterator<RepairSheet> iterator = repairSheets.iterator();
-  //     while(iterator.hasNext()){
-  //         RepairSheet repairSheet = iterator.next();
-  //         if(repairSheet.getType_of_maintenance().getType_of_maintenance_id()==3||repairSheet.getType_of_maintenance().getType_of_maintenance_id()==4
-  //                 ||repairSheet.getType_of_maintenance().getType_of_maintenance_id()==5){
-  //             iterator.remove();
-  //         }
-  //     }
+    @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
+    @RequestMapping(value = "/equipment/{equipmentId}", method = RequestMethod.GET,
+            produces = "application/json")
+    public @ResponseBody List<RepairSheet> getRepairByEquipmentId(@PathVariable("equipmentId") int equipmentId, Model model) {
+        List<RepairSheet> repairSheets = repairSheetService.getRepairSheetByEquipment(equipmentService.getEquipmentById(equipmentId));
+        Iterator<RepairSheet> iterator = repairSheets.iterator();
+        while(iterator.hasNext()){
+            RepairSheet repairSheet = iterator.next();
+            if(repairSheet.getType_of_maintenance().getType_of_maintenance_id()==3||repairSheet.getType_of_maintenance().getType_of_maintenance_id()==4
+                    ||repairSheet.getType_of_maintenance().getType_of_maintenance_id()==5){
+                iterator.remove();
+            }
+        }
 
-  //     return this.RepairSheetService.getRepairSheetByEquipment(this.EquipmentService.getEquipmentById(equipmentId));
-  // }
+        return repairSheetService.getRepairSheetByEquipment(equipmentService.getEquipmentById(equipmentId));
+    }
 
-   // @ApiOperation(value = "getAllRepairSheet", notes = "Get all equipments")
-   // @RequestMapping(value = "/all", method = RequestMethod.GET,
-   //         produces = "application/json")
-   // public @ResponseBody List<RepairSheet> getAllRepair(Model model) {
-   //     return this.RepairSheetService.listRepairSheets();
-   // }
-//
-  //  @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-  //  @RequestMapping(value = "/{repairId}", method = RequestMethod.POST,
-  //          produces = "application/json")
-  //  public void updateStatusRepair(@PathVariable("repairId") int repairId,
-  //                                 @RequestParam String date,
-  //                                 @RequestParam int status,
-  //                                 @RequestParam String description, Model model) throws ParseException  {
-  //      RepairSheet repairSheet= this.RepairSheetService.getRepairSheetById(repairId);
-  //      Status statusNew = this.StatusService.getStatusById(status);
-  //      repairSheet.setStatus(statusNew);
-  //      repairSheet.setDescription(description);
-  //      SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
-  //      Date newDate = formatter.parse(date);
-  //      if(status==5){
-  //          repairSheet.setEnd_date(newDate);
-  //      }
-  //      else {
-  //          repairSheet.setConfirm_date(newDate);
-  //      }
-  //      this.RepairSheetService.updateRepairSheet(repairSheet);
-//
-  //      changeStatusOfEquipment(repairSheet.getEquipment(),statusNew, repairSheet.getType_of_maintenance());
-  //  }
-//
-  //  @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-  //  @RequestMapping(value = "/confirm/{repairId}", method = RequestMethod.POST,
-  //          produces = "application/json")
-  //  public void confirmRepair(@PathVariable("repairId") int repairId,
-  //                            @RequestParam int status,
-  //                            @RequestParam String description, Model model) throws ParseException  {
-  //      RepairSheet repairSheet= this.RepairSheetService.getRepairSheetById(repairId);
-  //      Status statusNew = this.StatusService.getStatusById(status);
-  //      repairSheet.setStatus(statusNew);
-  //      if (description.length()!=0) {
-  //          repairSheet.addDescription(description);
-  //      }
-  //      this.RepairSheetService.updateRepairSheet(repairSheet);
-  //  }
+    @ApiOperation(value = "getAllRepairSheet", notes = "Get all equipments")
+    @RequestMapping(value = "/all", method = RequestMethod.GET,
+            produces = "application/json")
+    public @ResponseBody List<RepairSheet> getAllRepair(Model model) {
+        return repairSheetService.getAllRepairSheets();
+    }
 
+    @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
+    @RequestMapping(value = "/repair/{repairId}", method = RequestMethod.POST,
+            produces = "application/json")
+    public void updateStatusRepair(@PathVariable("repairId") int repairId,
+                                   @RequestParam String date,
+                                   @RequestParam int status,
+                                   @RequestParam String description, Model model) throws ParseException  {
+        RepairSheet repairSheet= repairSheetService.getRepairSheetById(repairId);
+        Status statusNew = statusService.getStatusById(status);
+        repairSheet.setStatus(statusNew);
+        repairSheet.setDescription(description);
+        SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
+        Date newDate = formatter.parse(date);
+        if(status==5){
+            repairSheet.setEnd_date(newDate);
+        }
+        else {
+            repairSheet.setConfirm_date(newDate);
+        }
+        repairSheetService.updateRepairSheet(repairSheet);
+
+        changeStatusOfEquipment(repairSheet.getEquipment(),statusNew, repairSheet.getType_of_maintenance());
+    }
+
+    @ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
+    @RequestMapping(value = "/confirm/{repairId}", method = RequestMethod.POST,
+            produces = "application/json")
+    public void confirmRepair(@PathVariable("repairId") int repairId,
+                              @RequestParam int status,
+                              @RequestParam String description, Model model) throws ParseException  {
+        RepairSheet repairSheet= repairSheetService.getRepairSheetById(repairId);
+        Status statusNew = statusService.getStatusById(status);
+        repairSheet.setStatus(statusNew);
+        if (description.length()!=0) {
+            repairSheet.addDescription(description);
+        }
+        repairSheetService.updateRepairSheet(repairSheet);
+    }
 
 }
