+++ src/main/java/com/diplom/sptor/web/ServiceController.java
@@ -2,7 +2,6 @@ package com.diplom.sptor.web;
 
 import com.diplom.sptor.domain.*;
 import com.diplom.sptor.model.UserFormModel;
-import com.diplom.sptor.repository.UserRepository;
 import com.diplom.sptor.service.*;
 import com.wordnik.swagger.annotations.Api;
 import com.wordnik.swagger.annotations.ApiOperation;
@@ -59,7 +58,7 @@ public class ServiceController {
 	SpareService spareService;
 
 	@Autowired
-	ComponentManager componentManager;
+	ComponentService componentService;
 
 	@Autowired
 	EquipmentService equipmentService;
@@ -170,14 +169,14 @@ public class ServiceController {
 		Equipment equipment = equipmentService.getEquipmentById(equipmentId);
 		TypeOfEquipment typeOfEquipment = equipment.getTypeOfEquipment();
 		//TypeOfEquipment typeOfEquipment=  TypeOfEquipmentService.getTypeById(typeId);
-		return componentManager.getComponentByType(typeOfEquipment);
+		return componentService.getComponentByType(typeOfEquipment);
 	}
 
 	@RequestMapping(value = "/components/{componentId}/equipment/{equipmentId}", method = RequestMethod.GET,
 			produces = "application/json")
 	public @ResponseBody List<RepairSheet> getTypeOfMaintenanceByComponents(@PathVariable(value = "componentId")int componentId, @PathVariable(value = "equipmentId")int equipmentId,  Model model) {
 		Equipment equipment = equipmentService.getEquipmentById(equipmentId);
-		Components components = componentManager.getComponentById(componentId);
+		Components components = componentService.getComponentById(componentId);
 		return repairSheetService.getTypeOfMaintenanceOfRepairByEquipmentAndComponents(equipment, components);
 	}
 
@@ -216,172 +215,6 @@ public class ServiceController {
 		this.subdivisionService.addSubdivision(subdivision);
 		return "Added successfully";
 
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all equipments")
-	@RequestMapping(value = "/repair", method = RequestMethod.GET,
-			produces = "application/json")
-	public String RepairStart(Model model) {
-		model.addAttribute("repairSheet", new RepairSheet());
-		model.addAttribute("listRepair", repairSheetService.getAllRepairSheets());
-		model.addAttribute("status_one", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(1)).size());
-		model.addAttribute("status_two", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(2)).size() + repairSheetService.getRepairSheetByStatus(statusService.getStatusById(3)).size());
-		model.addAttribute("status_three", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(4)).size());
-		model.addAttribute("status_four", repairSheetService.getRepairSheetByStatus(statusService.getStatusById(5)).size());
-		model.addAttribute("status_all", repairSheetService.getAllRepairSheets().size());
-		model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
-		model.addAttribute("components", componentManager.getAllComponents());
-		model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
-				userService.getUserBySso(getPrincipal()).getFirst_name());
-		model.addAttribute("listEquipments", equipmentService.getAllEquipment());
-		model.addAttribute("listTypeOfMaintenance", typeOfMaintenanceService.getAllTypes());
-		Status status1 = statusService.getStatusById(1);
-		Status status2 = statusService.getStatusById(2);
-		model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
-		model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
-		model.addAttribute("listStatus", repairSheetService.getAllRepairSheets());
-		return "repair";
-	}
-
-
-
-	@RequestMapping(value = "/repair/add", method = RequestMethod.POST)
-	public String addRepairSheet(
-			@RequestParam int subdivision_id,
-			@RequestParam int equipment_id,
-			@RequestParam int components,
-			@RequestParam int type_of_maintenance_id,
-			@RequestParam String start_date,
-			@RequestParam String description
-	) throws ParseException {
-		SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
-		Date start = formatter.parse(start_date);
-		Date end = formatter.parse(start_date);
-		DateTime dateTime = new DateTime(end);
-		DateTimeFormatter dtf = DateTimeFormat.forPattern("mm-dd-yyyy");
-		DateTime end_date = dateTime.plusDays(typeOfMaintenanceService.getTypeById(type_of_maintenance_id).getDuration());
-		Status status = statusService.getStatusById(1);
-		User resp_for_delivery = userService.getUserById(1);
-		User resp_for_reception = userService.getUserById(1);
-		Equipment equipment = equipmentService.getEquipmentById(equipment_id);
-		Components components1 = componentManager.getComponentById(components);
-		Subdivisions subdivisions = subdivisionService.getSubdivisionById(subdivision_id);
-		TypeOfMaintenance typeOfMaintenance = typeOfMaintenanceService.getTypeById(type_of_maintenance_id);
-		int sheet_number = equipment_id + 33000;
-		int warranty_period = type_of_maintenance_id;
-		repairSheetService.addRepairSheet(new RepairSheet(typeOfMaintenance, equipment,
-				components1, subdivisions, start, end_date.toDate(), sheet_number, warranty_period,
-				resp_for_delivery, resp_for_reception, description, status));
-
-		deleteComponentInStock(components1);
-		changeStatusOfEquipment(equipment, status, typeOfMaintenance);
-
-		return "redirect:/repair";
-
-	}
-
-	public void deleteComponentInStock(Components components){
-		Spares spares =  spareService.getSpareById(components.getSpare().getSpare_id());
-		spares.setAmount_in_stock(spares.getAmount_in_stock() - 1);
-		//this.SpareService.updateSpare(spares);
-	}
-
-	public void changeStatusOfEquipment(Equipment equipment, Status status, TypeOfMaintenance typeOfMaintenance){
-		int type = typeOfMaintenance.getType_of_maintenance_id();
-		switch (status.getStatus_id()){
-			case 1:
-				if(type==1||type==2||type==6||type==7||type==8){
-					equipment.setStatus(statusOfEqService.getStatusById(2));
-					equipmentService.updateEquipment(equipment);
-				}
-				else if(type==3||type==4||type==5){
-					equipment.setStatus(statusOfEqService.getStatusById(3));
-					equipmentService.updateEquipment(equipment);
-				}
-				else{
-					equipment.setStatus(statusOfEqService.getStatusById(4));
-					equipmentService.updateEquipment(equipment);
-				}
-				break;
-			case 4:
-				equipment.setStatus(statusOfEqService.getStatusById(1));
-				equipmentService.updateEquipment(equipment);
-				break;
-			case 5:
-				equipment.setStatus(statusOfEqService.getStatusById(1));
-				equipmentService.updateEquipment(equipment);
-				break;
-		}
-	}
-
-	@ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-	@RequestMapping(value = "/repair/{repairId}", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody RepairSheet getRepairById(@PathVariable("repairId") int repairId, Model model) {
-		return repairSheetService.getRepairSheetById(repairId);
-	}
-
-	@ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-	@RequestMapping(value = "/repair/equipment/{equipmentId}", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody List<RepairSheet> getRepairByEquipmentId(@PathVariable("equipmentId") int equipmentId, Model model) {
-		List<RepairSheet> repairSheets = repairSheetService.getRepairSheetByEquipment(equipmentService.getEquipmentById(equipmentId));
-		Iterator<RepairSheet> iterator = repairSheets.iterator();
-		while(iterator.hasNext()){
-			RepairSheet repairSheet = iterator.next();
-			if(repairSheet.getType_of_maintenance().getType_of_maintenance_id()==3||repairSheet.getType_of_maintenance().getType_of_maintenance_id()==4
-					||repairSheet.getType_of_maintenance().getType_of_maintenance_id()==5){
-				iterator.remove();
-			}
-		}
-
-		return repairSheetService.getRepairSheetByEquipment(equipmentService.getEquipmentById(equipmentId));
-	}
-
-	@ApiOperation(value = "getAllRepairSheet", notes = "Get all equipments")
-	@RequestMapping(value = "/repair/all", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody List<RepairSheet> getAllRepair(Model model) {
-		return repairSheetService.getAllRepairSheets();
-	}
-
-	@ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-	@RequestMapping(value = "/repair/{repairId}", method = RequestMethod.POST,
-			produces = "application/json")
-	public void updateStatusRepair(@PathVariable("repairId") int repairId,
-								   						@RequestParam String date,
-														@RequestParam int status,
-														@RequestParam String description, Model model) throws ParseException  {
-		RepairSheet repairSheet= repairSheetService.getRepairSheetById(repairId);
-		Status statusNew = statusService.getStatusById(status);
-		repairSheet.setStatus(statusNew);
-		repairSheet.setDescription(description);
-		SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
-		Date newDate = formatter.parse(date);
-		if(status==5){
-			repairSheet.setEnd_date(newDate);
-		}
-		else {
-			repairSheet.setConfirm_date(newDate);
-		}
-		repairSheetService.updateRepairSheet(repairSheet);
-
-		changeStatusOfEquipment(repairSheet.getEquipment(),statusNew, repairSheet.getType_of_maintenance());
-	}
-
-	@ApiOperation(value = "getRepairSheetById", notes = "Get all equipments")
-	@RequestMapping(value = "/repair/confirm/{repairId}", method = RequestMethod.POST,
-			produces = "application/json")
-	public void confirmRepair(@PathVariable("repairId") int repairId,
-								   @RequestParam int status,
-								   @RequestParam String description, Model model) throws ParseException  {
-		RepairSheet repairSheet= repairSheetService.getRepairSheetById(repairId);
-		Status statusNew = statusService.getStatusById(status);
-		repairSheet.setStatus(statusNew);
-		if (description.length()!=0) {
-			repairSheet.addDescription(description);
-		}
-		repairSheetService.updateRepairSheet(repairSheet);
 	}
 
 	@RequestMapping(value = "/report/repair/act_in/{repairId}", method = RequestMethod.POST)
@@ -787,267 +620,4 @@ public class ServiceController {
 		technologicalCardService.deleteCard(id);
 		return "Deleted successfully";
 	}
-
-	@ApiOperation(value = "getUsers", notes = "Get all technological cards")
-	@RequestMapping(value = "/manual", method = RequestMethod.GET,
-			produces = "application/json")
-	public String getManual(Model model) {
-		model.addAttribute("userForm", new UserFormModel());
-		Status status1 = statusService.getStatusById(1);
-		Status status2 = statusService.getStatusById(2);
-		model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
-		model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
-		model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
-		model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
-				userService.getUserBySso(getPrincipal()).getFirst_name());
-		return "manual";
-	}
-
-
-	@ApiOperation(value = "getUsers", notes = "Get all technological cards")
-	@RequestMapping(value = "/users/all", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody List<User> getUsers(Model model) {
-		return userService.getUsers();
-	}
-
-	@ApiOperation(value = "getUsers", notes = "Get all technological cards")
-	@RequestMapping(value = "/users", method = RequestMethod.GET,
-			produces = "application/json")
-	public String getUsersView(Model model) {
-		model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
-		model.addAttribute("userForm", new UserFormModel());
-		model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
-				userService.getUserBySso(getPrincipal()).getFirst_name());
-		Status status1 = statusService.getStatusById(1);
-		Status status2 = statusService.getStatusById(2);
-		model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
-		model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
-		return "users";
-	}
-
-	@ApiOperation(value = "addUser", notes = "Get all technological cards")
-	@RequestMapping(value = "/users", method = RequestMethod.POST,
-			produces = "application/json")
-	public String addUsers(@RequestParam String ssoid,
-						   @RequestParam String first_name,
-						   @RequestParam String last_name,
-						   @RequestParam String password,
-						   @RequestParam String email) throws NoSuchAlgorithmException, UnsupportedEncodingException{
-		MessageDigest md = MessageDigest.getInstance("SHA-256");
-		md.update(password.getBytes("UTF-8"));
-		byte [] digest = md.digest();
-		User user = new User();
-		user.setFirst_name(first_name);
-		user.setLast_name(last_name);
-		user.setEmail(email);
-		user.setssoid(ssoid);
-		user.setPassword(String.format("%064x", new java.math.BigInteger(1, digest)));
-		userService.addUser(user);
-		return "manual";
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all technological cards")
-	@RequestMapping(value = "/equipment/all", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody List<Equipment> getEquipments(Model model) {
-		return equipmentService.getAllEquipment();
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all technological cards")
-	@RequestMapping(value = "/equipment", method = RequestMethod.GET,
-			produces = "application/json")
-	public String getEquipmentView(Model model) {
-		model.addAttribute("eqType", typeOfEquipmentService.getAllTypes());
-		model.addAttribute("eqSub", subdivisionService.getAllSubdivisions());
-		model.addAttribute("eqResp", this.userService.getUsers());
-		model.addAttribute("eqStatus", statusOfEqService.listStatus());
-		model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
-		model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
-				userService.getUserBySso(getPrincipal()).getFirst_name());
-		Status status1 = statusService.getStatusById(1);
-		Status status2 = statusService.getStatusById(2);
-		model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
-		model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
-		return "equipments";
-	}
-
-	@ApiOperation(value = "addEquipment", notes = "Get all technological cards")
-	@RequestMapping(value = "/equipment", method = RequestMethod.POST,
-			produces = "application/json")
-	public String addEquipment(@RequestParam String equipment_name,
-						   @RequestParam int type_of_equipment_id,
-						   @RequestParam int subdivision_id,
-						   @RequestParam int inventory_number,
-						   @RequestParam int graduation_year,
-						   @RequestParam String producer_of_equipment,
-						   @RequestParam double working_hours,
-						   @RequestParam int status,
-						   @RequestParam String description) throws ParseException {
-		SimpleDateFormat formatter = new SimpleDateFormat("yyyy");
-		Date date = formatter.parse(String.valueOf(graduation_year));
-		Equipment equipment = new Equipment();
-		equipment.setEquipmentName(equipment_name);
-		equipment.setTypeOfEquipment(typeOfEquipmentService.getTypeById(type_of_equipment_id));
-		equipment.setSubdivision(subdivisionService.getSubdivisionById(subdivision_id));
-		equipment.setInventoryNumber(inventory_number);
-		equipment.setGraduationYear(date);
-		equipment.setProducerOfEquipment(producer_of_equipment);
-		equipment.setWorkingHours(working_hours);
-		equipment.setStatus(statusOfEqService.getStatusById(status));
-		equipment.setDescription(description);
-		equipmentService.addEquipment(equipment);
-		return "manual";
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all technological cards")
-	@RequestMapping(value = "/spare/all", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody List<Spares> getSpare(Model model) {
-		return spareService.getAllSpares();
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all technological cards")
-	@RequestMapping(value = "/spare", method = RequestMethod.GET,
-			produces = "application/json")
-	public String getSpareView(Model model) {
-		model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
-		model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
-				userService.getUserBySso(getPrincipal()).getFirst_name());
-		Status status1 = statusService.getStatusById(1);
-		Status status2 = statusService.getStatusById(2);
-		model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
-		model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
-		return "spare";
-	}
-
-	@ApiOperation(value = "addEquipment", notes = "Get all technological cards")
-	@RequestMapping(value = "/spare", method = RequestMethod.POST,
-			produces = "application/json")
-	public String addSpare(@RequestParam String spare_name,
-							   @RequestParam String spare_producer,
-							   @RequestParam double price,
-						       @RequestParam int amount_in_stock,
-							   @RequestParam String description) throws ParseException {
-		Spares spare = new Spares();
-		spare.setSpare_name(spare_name);
-		spare.setSpare_producer(spare_producer);
-		spare.setPrice(price);
-		spare.setAmount_in_stock(amount_in_stock);
-		spare.setDescription(description);
-		spareService.saveSpare(spare);
-		return "manual";
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all technological cards")
-	@RequestMapping(value = "/subdivision/all", method = RequestMethod.GET,
-			produces = "application/json")
-	public @ResponseBody List<Subdivisions> getSubdivision(Model model) {
-		return subdivisionService.getAllSubdivisions();
-	}
-
-	@ApiOperation(value = "getEquipments", notes = "Get all technological cards")
-	@RequestMapping(value = "/subdivision", method = RequestMethod.GET,
-			produces = "application/json")
-	public String getSubdivisionView(Model model) {
-		model.addAttribute("subdivisions", subdivisionService.getAllSubdivisions());
-		model.addAttribute("user", userService.getUserBySso(getPrincipal()).getLast_name() + " " +
-				userService.getUserBySso(getPrincipal()).getFirst_name());
-		Status status1 = statusService.getStatusById(1);
-		Status status2 = statusService.getStatusById(2);
-		model.addAttribute("active_req", repairSheetService.getRepairSheetByStatus(status1).size());
-		model.addAttribute("confirm_req", repairSheetService.getRepairSheetByStatus(status2).size());
-		return "subdivisions";
-	}
-
-	@ApiOperation(value = "addEquipment", notes = "Get all technological cards")
-	@RequestMapping(value = "/subdivision", method = RequestMethod.POST,
-			produces = "application/json")
-	public String addSubdivision(@RequestParam String subdivision_name,
-						   @RequestParam String abbreviation,
-						   @RequestParam String responsible,
-						   @RequestParam String description) throws ParseException {
-		Subdivisions subdivisions = new Subdivisions();
-		subdivisions.setSubdivision_name(subdivision_name);
-		subdivisions.setAbbreviation(abbreviation);
-		subdivisions.setResponsible(responsible);
-		subdivisions.setDescription(description);
-		subdivisionService.addSubdivision(subdivisions);
-		return "manual";
-	}
-
-	@RequestMapping(value = "/type_of_equipment", method = RequestMethod.GET)
-	public String listTypes(Model model) {
-		model.addAttribute("type_of_equipment", new TypeOfEquipment());
-		model.addAttribute("listTypes", typeOfEquipmentService.getAllTypes());
-		return "type_of_equipment";
-	}
-
-	@RequestMapping(value= "/type_of_equipment/add", method = RequestMethod.POST)
-	public String addType(@ModelAttribute("type_of_equipment") TypeOfEquipment typeOfEquipment, Model model){
-		if(typeOfEquipment.getType_of_equipment_id() == 0){
-			//new person, add it
-			typeOfEquipmentService.addType(typeOfEquipment);
-		}else{
-			//existing person, call update
-			typeOfEquipmentService.updateType(typeOfEquipment);
-		}
-		model.addAttribute("type_of_equipment", new TypeOfEquipment());
-		return "type_of_equipment";
-
-	}
-
-
-/*
-	@RequestMapping("type/edit/{id}")
-	public String editType(@PathVariable("id") int id, Model model){
-		model.addAttribute("type_of_equipment", this.TypeOfEquipmentService.getTypeById((id)));
-		model.addAttribute("listTypes", this.TypeOfEquipmentService.listType());
-
-		return "type_of_equipment";
-	}
-
-	@RequestMapping(value = "type/remove/{id}")
-	public String removeType(@PathVariable("id") int id){
-
-		this.TypeOfEquipmentService.deleteType(id);
-		return "redirect:/type_of_equipment";
-	}
-
-	@RequestMapping(value = "/spares", method = RequestMethod.GET)
-	public String listSpares(Model model) {
-		model.addAttribute("spare", new Spares());
-		model.addAttribute("listSpares", this.SpareService.listSpares());
-		return "spare";
-	}
-
-	@RequestMapping(value= "/spares/add", method = RequestMethod.POST)
-	public String addSpare(@ModelAttribute("spares") Spares spares){
-		if(spares.getSpare_id() == 0){
-			//new person, add it
-			this.SpareService.addSpare(spares);
-		}else{
-			//existing person, call update
-			this.SpareService.updateSpare(spares);
-		}
-
-		return "redirect:/spares";
-
-	}
-
-	@RequestMapping("spares/edit/{id}")
-	public String editSpare(@PathVariable("id") int id, Model model){
-		model.addAttribute("spare", this.SpareService.getSpareById((id)));
-		model.addAttribute("listSpares", this.SpareService.listSpares());
-
-		return "spare";
-	}
-
-	@RequestMapping(value = "spares/remove/{id}")
-	public String removeSpare(@PathVariable("id") int id){
-
-		this.SpareService.deleteSpare(id);
-		return "redirect:/spares";
-	}*/
-
 }
